Revert "Pro 38 finish game (#11)"
This reverts commit 5df3372e23327c03b9cd849a7a1c8662bb5b0ffa.
