--- colors.py
@@ -1,8 +1,0 @@
-# --- Константы ---
-COLOR_BACKGROUND = (30, 30, 30)
-COLOR_GRID_DEFAULT = (200, 200, 200)
-COLOR_GRID_HOVERED = (255, 255, 100)
-COLOR_GRID_ACTIVE_UNIT = (100, 200, 255)
-COLOR_GRID_OUTLINE = (0, 0, 0)
-COLOR_PATH_HOVER = (100, 255, 100)
-COLOR_PATH_ACTIVE = (100, 255, 100)
