--- commands/__init__.py
@@ -1,9 +1,0 @@
-# components/__init__.py
-
-from .attack import AttackCommand
-from .move import MoveCommand
-from .queued_attack import QueuedAttack
-
-__all__ = [
-    "AttackCommand", "MoveCommand", "QueuedAttack"
-]
