--- commands/attack.py
@@ -1,5 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class AttackCommand:
-    target_id: int  # цель для атаки
