--- commands/move.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class MoveCommand:
-    path: list  # список клеток [(q, r), ...]
-
