--- commands/queued_attack.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class QueuedAttack:
-    target_id: int  # цель для отложенной атаки
-
