--- components/__init__.py
@@ -1,22 +1,0 @@
-# components/__init__.py
-
-from .hex_position import HexPosition
-from .hovered import Hovered
-from .clickable import Clickable
-from .renderable import Renderable
-from .animation import Animation
-from .initiative import Initiative
-from .active_turn import ActiveTurn
-from .path import Path
-from .blocking_move import BlockingMove
-from .health import Health
-from .team import Team
-from .attack import Attack
-from .game_over import GameOver
-from .endgame_ui import EndgameUI
-
-__all__ = [
-    "HexPosition", "Hovered", "Clickable", "Renderable",
-    "Animation", "Initiative", "ActiveTurn", "Path", "BlockingMove",
-    "Health", "Team", "Attack", "GameOver", "EndgameUI"
-]
