--- components/active_turn.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class ActiveTurn:
-    pass
-
