--- components/blocking_move.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class BlockingMove:
-    pass
-
