--- components/clickable.py
@@ -1,2 +1,0 @@
-class Clickable:
-    pass
