--- components/endgame_ui.py
@@ -1,3 +1,0 @@
-class EndgameUI:
-    def __init__(self, window):
-        self.window = window
