--- components/game_over.py
@@ -1,3 +1,0 @@
-class GameOver:
-    def __init__(self):
-        self.message_shown = False
