--- components/health.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Health:
-    def __init__(self, max_value, value):
-        self.max_value = max_value
-        self.value = value
-
