--- components/hex_position.py
@@ -1,5 +1,0 @@
-class HexPosition:
-    def __init__(self, q, r):
-        self.q = q
-        self.r = r
-
