--- components/hovered.py
@@ -1,3 +1,0 @@
-class Hovered:
-    pass
-
