--- components/initiative.py
@@ -1,5 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Initiative:
-    value: int
