--- components/renderable.py
@@ -1,4 +1,0 @@
-class Renderable:
-    def __init__(self, color):
-        self.color = color
-
