--- components/team.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Team:
-    def __init__(self, name):
-        self.name = name
-
